Move Boundbox to its own file, remove non-Coords constructors
