+++ src/openrct2/paint/Supports.cpp
@@ -14,6 +14,7 @@
 #include "../sprites.h"
 #include "../util/Math.hpp"
 #include "../world/Surface.h"
+#include "Boundbox.h"
 #include "Paint.h"
 #include "tile_element/Paint.TileElement.h"
