+++ src/openrct2/world/Location.hpp
@@ -702,44 +702,6 @@ struct TileCoordsXYZD : public TileCoordsXYZ
     }
 };
 
-struct BoundBoxXY
-{
-    CoordsXY offset{};
-    CoordsXY length{};
-
-    constexpr BoundBoxXY() = default;
-    constexpr BoundBoxXY(CoordsXY _offset, CoordsXY _length)
-        : offset(_offset)
-        , length(_length)
-    {
-    }
-
-    constexpr BoundBoxXY(int32_t _offX, int32_t _offY, int32_t _lenX, int32_t _lenY)
-        : offset({ _offX, _offY })
-        , length({ _lenX, _lenY })
-    {
-    }
-};
-
-struct BoundBoxXYZ
-{
-    CoordsXYZ offset{};
-    CoordsXYZ length{};
-
-    constexpr BoundBoxXYZ() = default;
-    constexpr BoundBoxXYZ(CoordsXYZ _offset, CoordsXYZ _length)
-        : offset(_offset)
-        , length(_length)
-    {
-    }
-
-    constexpr BoundBoxXYZ(int32_t _offX, int32_t _offY, int32_t _offZ, int32_t _lenX, int32_t _lenY, int32_t _lenZ)
-        : offset({ _offX, _offY, _offZ })
-        , length({ _lenX, _lenY, _lenZ })
-    {
-    }
-};
-
 /**
  * Represents a range of the map using regular coordinates.
  */
